Move LangLibrary to impl package
